style corrections
