+++ test/AutomatedCarTests/SystemComponents/Packets/DebugPacketTestBase.cs
@@ -1,0 +1,15 @@
+﻿using AutomatedCar.SystemComponents.Packets;
+using Xunit;
+
+namespace Tests.SystemComponents.Packets
+{
+    public abstract class DebugPacketTestBase
+    {
+        public DebugPacket debugPacket;
+        protected DebugPacketTestBase()
+        {
+            debugPacket = new DebugPacket();
+        }
+
+    }
+}
