+++ test/AutomatedCarTests/SystemComponents/Packets/DebugPacketTests.cs
@@ -3,16 +3,6 @@ using Xunit;
 
 namespace Tests.SystemComponents.Packets
 {
-    public abstract class DebugPacketTestBase
-    {
-        public DebugPacket debugPacket;
-        protected DebugPacketTestBase()
-        {
-            debugPacket = new DebugPacket();
-        }
-
-    }
-
     public class DebugPacketTests : DebugPacketTestBase
     {
