+++ test/AutomatedCarTests/SystemComponents/Packets/HMIPacketTests.cs
@@ -3,16 +3,6 @@ using Xunit;
 
 namespace Tests.SystemComponents.Packets
 {
-    public abstract class HMIPacketTestBase
-    {
-        public HMIPacket hmiPacket;
-        protected HMIPacketTestBase()
-        {
-            hmiPacket = new HMIPacket();
-            hmiPacket.Sign = "";
-        }
-    }
-
     public class HMIPacketTests : HMIPacketTestBase
     {
